requirements.txt
# Core Framework
langgraph>=0.0.62
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.13

# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI/ML
openai>=1.6.0
tiktoken>=0.5.2

# Social Media APIs
facebook-sdk>=3.1.0
tweepy>=4.14.0
python-instagram>=1.3.2
TikTokApi>=5.2.4

# HTTP & API
httpx>=0.25.2
requests>=2.31.0
aiohttp>=3.9.1

# Database & Storage
faiss-cpu>=1.7.4
sqlite3
sqlalchemy>=2.0.23
alembic>=1.13.1

# Vector Store & Memory
chromadb>=0.4.18
sentence-transformers>=2.2.2

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.1
python-multipart>=0.0.6

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
schedule>=1.2.0
pillow>=10.1.0
jinja2>=3.1.2

# Development & Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1

# Logging & Monitoring
structlog>=23.2.0
prometheus-client>=0.19.0

# Security
cryptography>=41.0.8
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# MCP Framework
model-context-protocol>=0.1.0

# GoHighLevel Integration
requests-oauthlib>=1.3.1

# Async Support
asyncio
aiofiles>=23.2.1 